feat: Adds component loss database tests.
