+++ Tests/DatabaseClientTests/Helpers.swift
@@ -53,3 +53,15 @@ func withTestUser(
     try await operation(user)
   }
 }
+
+extension Project.Create {
+
+  static let mock = Self(
+    name: "Testy McTestface",
+    streetAddress: "1234 Sesame St",
+    city: "Nowhere",
+    state: "MN",
+    zipCode: "55555",
+    sensibleHeatRatio: 0.83
+  )
+}
